rename weekly report action
