+++ src/ducks/Watchlists/Actions/WeeklyReport/EmailImage.js
