+++ src/ducks/Watchlists/Actions/WeeklyReport/Trigger.module.scss
