+++ src/ducks/Watchlists/Actions/WeeklyReport/index.module.scss
