+++ src/epics/rootEpics.js
@@ -23,7 +23,7 @@ import {
   addAssetToWatchlistEpic,
   editAssetsInWatchlistEpic
 } from './../ducks/Watchlists/Actions/Edit/editAssetsInWatchlistEpic'
-import { watchlistToggleMonitoringEpic } from '../ducks/Watchlists/Actions/WatchlistWeeklyReport/watchlistToggleMonitoringEpic'
+import { watchlistToggleMonitoringEpic } from '../ducks/Watchlists/Actions/WeeklyReport/watchlistToggleMonitoringEpic'
 import {
   changeColumnsSettingsEpic,
   saveWatchlistsSettingsAfterLaunch
