+++ src/pages/Account/EmailSetting.js
@@ -4,7 +4,7 @@ import { compose } from 'recompose'
 import { graphql } from 'react-apollo'
 import { showNotification } from '../../actions/rootActions'
 import EditableInputSetting from './EditableInputSetting'
-import EditableInput from '../../ducks/Watchlists/Actions/WatchlistWeeklyReport/EditableInput'
+import EditableInput from '../../ducks/Watchlists/Actions/WeeklyReport/EditableInput'
 import { USER_EMAIL_CHANGE } from '../../actions/types'
 import { connect } from 'react-redux'
 import styles from './userName/UsernameSettings.module.scss'
